Apply code review suggestions part 2
