+++ src/main/java/bisq/daonode/DaoNode.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.daoNode;
+package bisq.daonode;
 
 
 import bisq.core.app.TorSetup;
